# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SP MLM Platform API",
        "description": "API Documentation for Mobile App and Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main Server"
        }
    ],
    "paths": {
...
        "/api/user/withdraw/config": {
            "get": {
                "tags": [
                    "User Withdrawal"
                ],
                "summary": "Get limits and bound accounts",
                "operationId": "e5ea9fe88c4ea2021d2b6e60cf561952",
                "responses": {
                    "200": {
                        "description": "Withdrawal limits and accounts fetched"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/withdraw": {
            "post": {
                "tags": [
                    "User Withdrawal"
                ],
                "summary": "Submit a Withdrawal Request",
                "operationId": "72b7b510ef14f0be146c1d9459fca02e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "wallet_type",
                                    "method"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "wallet_type": {
                                        "type": "string",
                                        "enum": [
                                            "personal",
                                            "income"
                                        ],
                                        "example": "income"
                                    },
                                    "method": {
                                        "type": "string",
                                        "enum": [
                                            "bank",
                                            "crypto"
                                        ],
                                        "example": "bank"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Withdrawal request submitted successfully"
                    },
                    "400": {
                        "description": "Business logic error (Limits, Balance etc.)"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/withdraw/history": {
            "get": {
                "tags": [
                    "User Withdrawal"
                ],
                "summary": "Get withdrawal history",
                "operationId": "c2556e3bf38b44fa6aa505eff20307e6",
                "responses": {
                    "200": {
                        "description": "History fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login with Mobile and Password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}